+++ cip25/rust/Cargo.toml
@@ -9,10 +9,10 @@ crate-type = ["cdylib", "rlib"]
 
 [dependencies]
 cml-core = { "path" = "../../core/rust" }
+cml-crypto = { "path" = "../../crypto/rust" }
+cml-chain = { "path" = "../../chain/rust" }
 cbor_event = "2.2.0"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0.57"
 schemars = "0.8.8"
-
-[dev-dependencies]
 hex = "0.4.0"
