+++ cip25/rust/src/serialization.rs
@@ -1,3 +1,5 @@
+use crate::utils::LabelMetadata;
+
 use super::*;
 pub use cml_core::{error::*, serialization::*};
 use std::io::{Seek, SeekFrom};
@@ -174,342 +176,6 @@ impl Deserialize for FilesDetails {
             })
         })()
         .map_err(|e| e.annotate("FilesDetails"))
-    }
-}
-
-impl cbor_event::se::Serialize for LabelMetadata {
-    fn serialize<'se, W: Write>(
-        &self,
-        serializer: &'se mut Serializer<W>,
-    ) -> cbor_event::Result<&'se mut Serializer<W>> {
-        match self {
-            LabelMetadata::LabelMetadataV1(label_metadata_v1) => {
-                serializer.write_map(cbor_event::Len::Len(label_metadata_v1.len() as u64))?;
-                for (key, value) in label_metadata_v1.iter() {
-                    key.serialize(serializer)?;
-                    serializer.write_map(cbor_event::Len::Len(value.len() as u64))?;
-                    for (key, value) in value.iter() {
-                        key.serialize(serializer)?;
-                        value.serialize(serializer)?;
-                    }
-                }
-                Ok(serializer)
-            }
-            LabelMetadata::LabelMetadataV2(label_metadata_v2) => {
-                label_metadata_v2.serialize(serializer)
-            }
-        }
-    }
-}
-
-impl Deserialize for LabelMetadata {
-    fn deserialize<R: BufRead + Seek>(raw: &mut Deserializer<R>) -> Result<Self, DeserializeError> {
-        (|| -> Result<_, DeserializeError> {
-            let initial_position = raw.as_mut_ref().seek(SeekFrom::Current(0)).unwrap();
-            match (|raw: &mut Deserializer<_>| -> Result<_, DeserializeError> {
-                let mut label_metadata_v1_table = BTreeMap::new();
-                let mut label_metadata_v1_table_len = 0;
-                let label_metadata_v1_len = raw.map()?;
-                while match label_metadata_v1_len {
-                    cbor_event::Len::Len(n) => label_metadata_v1_table_len < n as usize,
-                    cbor_event::Len::Indefinite => true,
-                } {
-                    match raw.cbor_type()? {
-                        cbor_event::Type::Text => {
-                            let label_metadata_v1_key = PolicyIdV1::deserialize(raw)?;
-                            let mut label_metadata_v1_value_table = BTreeMap::new();
-                            let mut label_metadata_v1_value_table_len = 0;
-                            let label_metadata_v1_value_len = raw.map()?;
-                            while match label_metadata_v1_value_len {
-                                cbor_event::Len::Len(n) => {
-                                    label_metadata_v1_value_table_len < n as usize
-                                }
-                                cbor_event::Len::Indefinite => true,
-                            } {
-                                match raw.cbor_type()? {
-                                    cbor_event::Type::Text => {
-                                        let label_metadata_v1_value_key =
-                                            AssetNameV1::deserialize(raw)?;
-                                        let label_metadata_v1_value_value =
-                                            MetadataDetails::deserialize(raw)?;
-                                        if label_metadata_v1_value_table
-                                            .insert(
-                                                label_metadata_v1_value_key.clone(),
-                                                label_metadata_v1_value_value,
-                                            )
-                                            .is_some()
-                                        {
-                                            return Err(DeserializeFailure::DuplicateKey(
-                                                Key::Str(String::from(
-                                                    "some complicated/unsupported type",
-                                                )),
-                                            )
-                                            .into());
-                                        }
-                                        label_metadata_v1_value_table_len += 1;
-                                    }
-                                    cbor_event::Type::Special => {
-                                        assert_eq!(raw.special()?, CBORSpecial::Break);
-                                        break;
-                                    }
-                                    _other_type => {
-                                        // we still need to read the data to move on to the CBOR after it
-                                        let _other_key =
-                                            cml_core::metadata::TransactionMetadatum::deserialize(
-                                                raw,
-                                            )?;
-                                        let _other_value =
-                                            cml_core::metadata::TransactionMetadatum::deserialize(
-                                                raw,
-                                            )?;
-                                        label_metadata_v1_value_table_len += 1;
-                                    }
-                                }
-                            }
-                            let label_metadata_v1_value = label_metadata_v1_value_table;
-                            if label_metadata_v1_table
-                                .insert(label_metadata_v1_key.clone(), label_metadata_v1_value)
-                                .is_some()
-                            {
-                                return Err(DeserializeFailure::DuplicateKey(Key::Str(
-                                    String::from("some complicated/unsupported type"),
-                                ))
-                                .into());
-                            }
-                            label_metadata_v1_table_len += 1;
-                        }
-                        cbor_event::Type::Special => {
-                            assert_eq!(raw.special()?, CBORSpecial::Break);
-                            break;
-                        }
-                        _other_type => {
-                            // we still need to read the data to move on to the CBOR after it
-                            let _other_key =
-                                cml_core::metadata::TransactionMetadatum::deserialize(raw)?;
-                            let _other_value =
-                                cml_core::metadata::TransactionMetadatum::deserialize(raw)?;
-                            label_metadata_v1_table_len += 1;
-                        }
-                    }
-                }
-                Ok(label_metadata_v1_table)
-            })(raw)
-            {
-                Ok(label_metadata_v1) => return Ok(Self::LabelMetadataV1(label_metadata_v1)),
-                Err(_) => raw
-                    .as_mut_ref()
-                    .seek(SeekFrom::Start(initial_position))
-                    .unwrap(),
-            };
-            match (|raw: &mut Deserializer<_>| -> Result<_, DeserializeError> {
-                Ok(LabelMetadataV2::deserialize(raw)?)
-            })(raw)
-            {
-                Ok(label_metadata_v2) => return Ok(Self::LabelMetadataV2(label_metadata_v2)),
-                Err(_) => raw
-                    .as_mut_ref()
-                    .seek(SeekFrom::Start(initial_position))
-                    .unwrap(),
-            };
-            Err(DeserializeError::new(
-                "LabelMetadata",
-                DeserializeFailure::NoVariantMatched.into(),
-            ))
-        })()
-        .map_err(|e| e.annotate("LabelMetadata"))
-    }
-}
-
-impl cbor_event::se::Serialize for LabelMetadataV2 {
-    fn serialize<'se, W: Write>(
-        &self,
-        serializer: &'se mut Serializer<W>,
-    ) -> cbor_event::Result<&'se mut Serializer<W>> {
-        serializer.write_map(cbor_event::Len::Len(2))?;
-        serializer.write_text(&"data")?;
-        serializer.write_map(cbor_event::Len::Len(self.data.len() as u64))?;
-        for (key, value) in self.data.iter() {
-            key.serialize(serializer)?;
-            serializer.write_map(cbor_event::Len::Len(value.len() as u64))?;
-            for (key, value) in value.iter() {
-                key.serialize(serializer)?;
-                value.serialize(serializer)?;
-            }
-        }
-        serializer.write_text(&"version")?;
-        serializer.write_unsigned_integer(2u64)?;
-        Ok(serializer)
-    }
-}
-
-impl Deserialize for LabelMetadataV2 {
-    fn deserialize<R: BufRead + Seek>(raw: &mut Deserializer<R>) -> Result<Self, DeserializeError> {
-        (|| -> Result<_, DeserializeError> {
-            let len = raw.map()?;
-            let mut read_len = CBORReadLen::new(match len {
-                cbor_event::Len::Len(n) => cbor_event::LenSz::Len(n, cbor_event::Sz::canonical(n)),
-                cbor_event::Len::Indefinite => cbor_event::LenSz::Indefinite,
-            });
-            read_len.read_elems(2)?;
-            let mut data = None;
-            let mut version_present = false;
-            let mut read = 0;
-            while match len {
-                cbor_event::Len::Len(n) => read < n as usize,
-                cbor_event::Len::Indefinite => true,
-            } {
-                match raw.cbor_type()? {
-                    CBORType::Text => match raw.text()?.as_str() {
-                        "data" => {
-                            if data.is_some() {
-                                return Err(DeserializeFailure::DuplicateKey(Key::Str(
-                                    "data".into(),
-                                ))
-                                .into());
-                            }
-                            data = Some(
-                                (|| -> Result<_, DeserializeError> {
-                                    let mut data_table = BTreeMap::new();
-                                    let data_len = raw.map()?;
-                                    let mut data_table_len = 0;
-                                    while match data_len {
-                                        cbor_event::Len::Len(n) => data_table_len < n as usize,
-                                        cbor_event::Len::Indefinite => true,
-                                    } {
-                                        match raw.cbor_type()? {
-                                            cbor_event::Type::Bytes => {
-                                                let data_key = PolicyIdV2::deserialize(raw)?;
-                                                let mut data_value_table_len = 0;
-                                                let mut data_value_table = BTreeMap::new();
-                                                let data_value_len = raw.map()?;
-                                                while match data_value_len {
-                                                    cbor_event::Len::Len(n) => data_value_table_len < n as usize,
-                                                    cbor_event::Len::Indefinite => true,
-                                                } {
-                                                    match raw.cbor_type()? {
-                                                        cbor_event::Type::Bytes => {
-                                                            let data_value_key = AssetNameV2::deserialize(raw)?;
-                                                            let data_value_value =
-                                                                MetadataDetails::deserialize(raw)?;
-                                                            if data_value_table
-                                                                .insert(data_value_key.clone(), data_value_value)
-                                                                .is_some()
-                                                            {
-                                                                return Err(DeserializeFailure::DuplicateKey(
-                                                                    Key::Str(String::from(
-                                                                        "some complicated/unsupported type",
-                                                                    )),
-                                                                )
-                                                                .into());
-                                                            }
-                                                            data_value_table_len += 1;
-                                                        },
-                                                        cbor_event::Type::Special => {
-                                                            assert_eq!(raw.special()?, CBORSpecial::Break);
-                                                            break;
-                                                        },
-                                                        _other_type => {
-                                                            // we still need to read the data to move on to the CBOR after it
-                                                            let _other_key = cml_core::metadata::TransactionMetadatum::deserialize(raw)?;
-                                                            let _other_value = cml_core::metadata::TransactionMetadatum::deserialize(raw)?;
-                                                            data_value_table_len += 1;
-                                                        },
-                                                    }
-                                                }
-                                                let data_value = data_value_table;
-                                                if data_table.insert(data_key.clone(), data_value).is_some()
-                                                {
-                                                    return Err(DeserializeFailure::DuplicateKey(
-                                                        Key::Str(String::from(
-                                                            "some complicated/unsupported type",
-                                                        )),
-                                                    )
-                                                    .into());
-                                                }
-                                                data_table_len += 1;
-                                            },
-                                            cbor_event::Type::Special => {
-                                                assert_eq!(raw.special()?, CBORSpecial::Break);
-                                                break;
-                                            },
-                                            _other_type => {
-                                                // we still need to read the data to move on to the CBOR after it
-                                                let _other_key = cml_core::metadata::TransactionMetadatum::deserialize(raw)?;
-                                                let _other_value = cml_core::metadata::TransactionMetadatum::deserialize(raw)?;
-                                                data_table_len += 1;
-                                            },
-                                        }
-                                    }
-                                    Ok(data_table)
-                                })()
-                                .map_err(|e| e.annotate("data"))?,
-                            );
-                        }
-                        "version" => {
-                            if version_present {
-                                return Err(DeserializeFailure::DuplicateKey(Key::Str(
-                                    "version".into(),
-                                ))
-                                .into());
-                            }
-                            version_present = (|| -> Result<_, DeserializeError> {
-                                let version_value = raw.unsigned_integer()?;
-                                if version_value != 2 {
-                                    return Err(DeserializeFailure::FixedValueMismatch {
-                                        found: Key::Uint(version_value),
-                                        expected: Key::Uint(2),
-                                    }
-                                    .into());
-                                }
-                                Ok(true)
-                            })()
-                            .map_err(|e| e.annotate("version"))?;
-                        }
-                        _unknown_key => {
-                            // CIP-25 allows permissive parsing
-                            read_len.read_elems(1)?;
-                            // we still need to read the data to move on to the CBOR after it
-                            let _other_metadatum = cml_core::metadata::TransactionMetadatum::deserialize(raw)?;
-                        }
-                    },
-                    CBORType::Special => match len {
-                        cbor_event::Len::Len(_) => {
-                            return Err(DeserializeFailure::BreakInDefiniteLen.into())
-                        }
-                        cbor_event::Len::Indefinite => match raw.special()? {
-                            CBORSpecial::Break => break,
-                            _ => return Err(DeserializeFailure::EndingBreakMissing.into()),
-                        },
-                    },
-                    _other_type => {
-                        // CIP-25 allows permissive parsing
-                        read_len.read_elems(1)?;
-                        // we still need to read the data to move on to the CBOR after it
-                        let _other_key = cml_core::metadata::TransactionMetadatum::deserialize(raw)?;
-                        let _other_value = cml_core::metadata::TransactionMetadatum::deserialize(raw)?;
-                    }
-                }
-                read += 1;
-            }
-            let data = match data {
-                Some(x) => x,
-                None => {
-                    return Err(
-                        DeserializeFailure::MandatoryFieldMissing(Key::Str(String::from("data")))
-                            .into(),
-                    )
-                }
-            };
-            if !version_present {
-                return Err(
-                    DeserializeFailure::MandatoryFieldMissing(Key::Str(String::from("version")))
-                        .into(),
-                );
-            }
-            ();
-            Ok(Self { data })
-        })()
-        .map_err(|e| e.annotate("LabelMetadataV2"))
     }
 }
